+++ plugins/apache/apache.go
@@ -19,8 +19,8 @@ type Apache struct {
 }
 
 var sampleConfig = `
-# An array of Apache status URI to gather stats.
-urls = ["http://localhost/server-status?auto"]
+	# An array of Apache status URI to gather stats.
+	urls = ["http://localhost/server-status?auto"]
 `
 
 func (n *Apache) SampleConfig() string {